+++ src/components/App/ModalApp.tsx
@@ -106,7 +106,7 @@ class ModalApp extends Component<Props> {
   }
 
   onInvalidCustomApiException = (callbackName: string) => {
-    const message = `CustomApiException: ${callbackName} must be a function that returns a promise for useCustomApiRequests to work properly.`
+    const message = `CustomApiException: ${callbackName} must be a function that returns a promise for useCustomizedApiRequests to work properly.`
     this.events.emit('error', { type: 'exception', message })
     Tracker.trackException(message)
   }
@@ -318,7 +318,7 @@ class ModalApp extends Component<Props> {
       actions.setDecoupleFromAPI(true)
     } else {
       actions.setDecoupleFromAPI(false)
-      this.onInvalidEnterpriseFeatureException('useCustomApiRequests')
+      this.onInvalidEnterpriseFeatureException('useCustomizedApiRequests')
     }
   }