+++ src/components/Router/CrossDeviceMobileRouter.tsx
@@ -230,6 +230,13 @@ export default class CrossDeviceMobileRouter extends Component<
           this.props.actions.showLogoCobranding(enterpriseFeatures.logoCobrand)
         }
       }
+
+      if (
+        enterpriseFeatures.useCustomizedApiRequests &&
+        validEnterpriseFeatures?.useCustomizedApiRequests
+      ) {
+        this.props.actions.setDecoupleFromAPI(true)
+      }
     } else {
       this.props.actions.hideOnfidoLogo(false)
       this.props.actions.showCobranding(null)