+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 ### Added
 
 - Internal: Introduce `SdkOptionsProvider` component and `useSdkOptions()` hook for SDK options' single source of truth.
+- Public: Added cross-device support for useCustomizedApiRequests callbacks via customer hosting of SDK. Note - This is a premium enterprise feature.
 - Public: Added support for UI customizations in SDK configuration using `customUI` option. See README for details of supported customization options.
 - Internal: Add Woopra tracking for UI customization option usage.